requirements.txt
anyio==4.4.0
appscript==1.2.5
certifi==2024.7.4
charset-normalizer==3.3.2
cloudconvert==2.1.0
docx2pdf==0.1.8
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.2
python-telegram-bot==21.4
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.5
urllib3==2.2.2
wheel==0.44.0
